move tag to global components folder
