+++ site/src/pages/critria/[id].astro
@@ -4,9 +4,9 @@ import { LuArrowLeft, LuBookMarked } from "react-icons/lu";
 import Button from "@/components/Button";
 import Heading from "@/components/Heading";
 import TableOfContents from "@/components/TableOfContents";
+import Tag from "@/components/Tag";
 import EvidenceTable from "@/critria/EvidenceTable";
 import { tagOptions } from "@/data/tags";
-import Tag from "@/locus/Tag";
 import curations from "~/criTRia-curations.json";
 
 /** generate pages for each datum, paths based on id */
