+++ site/src/pages/loci/[id].astro
@@ -12,6 +12,7 @@ import LineChart from "@/components/LineChart";
 import Link from "@/components/Link";
 import ShowMoreLines from "@/components/ShowMoreLines";
 import TableOfContents from "@/components/TableOfContents";
+import Tag from "@/components/Tag";
 import { deriveLocus } from "@/data/derived";
 import { getIdentifiers } from "@/data/identifiers";
 import { getInheritance } from "@/data/inheritance";
@@ -20,7 +21,6 @@ import AdditionalLiterature from "@/locus/AdditionalLiterature";
 import Citation from "@/locus/Citation";
 import Cited from "@/locus/Cited";
 import Gnomad from "@/locus/Gnomad";
-import Tag from "@/locus/Tag";
 import { getJsonBlame } from "@/util/git";
 import curations from "~/criTRia-curations.json";
 import gnomad from "~/plots/gnomad.json";
